move handlers to handlers.go, clean up unused config options
