+++ bridge/config/config.go
@@ -87,7 +87,7 @@ type Protocol struct {
 	IgnoreMessages         string // all protocols
 	Jid                    string // xmpp
 	Label                  string // all protocols
-	Login                  string // mattermost, matrix, keybase
+	Login                  string // mattermost, matrix
 	MediaDownloadBlackList []string
 	MediaDownloadPath      string // Basically MediaServerUpload, but instead of uploading it, just write it to a file on the same server.
 	MediaDownloadSize      int    // all protocols
@@ -145,7 +145,6 @@ type Protocol struct {
 	VerboseJoinPart        bool       // IRC
 	WebhookBindAddress     string     // mattermost, slack
 	WebhookURL             string     // mattermost, slack
-	PaperKey               string     // keybase
 }
 
 type ChannelOptions struct {
